Address comments, update licenses and javadoc
