+++ semantic-reranker/src/main/java/org/opensearch/search/relevance/control/KendraSearchExtBuilder.java
@@ -1,3 +1,10 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
 package org.opensearch.search.relevance.control;
 
 import java.io.IOException;
