+++ semantic-reranker/src/main/java/org/opensearch/search/relevance/ranker/Ranker.java
@@ -1,0 +1,23 @@
+package org.opensearch.search.relevance.ranker;
+
+import org.opensearch.action.search.SearchRequest;
+import org.opensearch.search.SearchHits;
+import org.opensearch.search.relevance.preprocess.QueryParser.QueryParserResult;
+
+public interface Ranker {
+
+  /**
+   * Decide whether the input request is eligible for re-ranking
+   * @param request input request
+   * @return boolean decision on whether to re-rank
+   */
+  boolean shouldRerank(final SearchRequest request);
+
+  /**
+   * Rank hits based on the provided query
+   * @param hits hits to be re-ranked
+   * @param queryParserResult query text and document field name to be used by the ranker
+   * @return SearchHits ordered by score generated by ranker
+   */
+  SearchHits rank(final SearchHits hits, QueryParserResult queryParserResult);
+}
